MyPage creator components로 빼고 파일정리 완료
