--- lib/components/custom_form.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../size.dart';
-import '../validate.dart';
-import 'custom_form_field.dart';
-
