+++ lib/holder/MyPage/creator/components/creator_my_page_header.dart
@@ -1,11 +1,10 @@
-import 'package:final_project_team02/holder/MyPage/user/components/user_my_page_order_mileage.dart';
-import 'package:final_project_team02/holder/MyPage/user/user_my_page.dart';
 import 'package:flutter/material.dart';
 
-import '../../../../components/custom_button.dart';
+import '../../_components/my_page_custom_button.dart';
+import '../../_components/my_page_order_mileage.dart';
 import '../../../../theme.dart';
 
-class MyPageheader extends StatelessWidget {
+class CreatorMyPageheader extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(
@@ -13,13 +12,40 @@ class MyPageheader extends StatelessWidget {
       child: Column(
         children: [
           SizedBox(height: 30),
-          user_my_page_pic(),
+          _buildCreatorPic(),
           SizedBox(height: 20),
-          custom_button(title: "프로필 설정"),
+          MyPageCustomButton(title: "프로필 설정"),
           SizedBox(height: 20),
-          UserMyPageOrderMileage(orderId: 10, mileageId: 200),
+          MyPageOrderMileage(orderId: 10, mileageId: 200),
         ],
       ),
+    );
+  }
+
+  Widget _buildCreatorPic() {
+    return Row(
+      children: [
+        SizedBox(
+          width: 65,
+          height: 65,
+          child: ClipRRect(
+            borderRadius: BorderRadius.circular(32.5),
+            child: Image.network(
+              'https://picsum.photos/200/100', // :TODO 04 사진수정
+              fit: BoxFit.cover,
+            ),
+          ),
+        ),
+        SizedBox(width: 16),
+        Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text('회원 이름', style: textTheme().displayMedium), // :TODO 04수정
+            Text('180cm • 70kg • 직장인',
+                style: textTheme().bodyMedium), // :TODO 04수정
+          ],
+        ),
+      ],
     );
   }
 }
