+++ lib/holder/MyPage/creator/creator_my_page.dart
@@ -1,8 +1,10 @@
+import 'package:final_project_team02/holder/MyPage/creator/components/creator_my_page_header.dart';
 import 'package:flutter/material.dart';
 
 import '../../../components/custom_app_bar.dart';
 import '../user/components/user_my_page_bottom.dart';
 import '../user/components/user_my_page_header.dart';
+import 'components/creator_my_page_bottom.dart';
 
 class CreatorMyPage extends StatelessWidget {
   @override
@@ -15,11 +17,11 @@ class CreatorMyPage extends StatelessWidget {
           return [
             SliverToBoxAdapter(
               //프로필 설정 버튼까지
-              child: MyPageheader(),
+              child: CreatorMyPageheader(),
             ),
           ];
         },
-        body: MyPageBody(),
+        body: CreatorMyPageBottom(),
       ),
     );
   }
