+++ lib/holder/MyPage/user/components/user_my_page_bottom.dart
@@ -1,4 +1,4 @@
-import 'package:final_project_team02/holder/MyPage/user/components/user_apply_button.dart';
+import 'package:final_project_team02/holder/MyPage/user/components/use_my_pager_apply_button.dart';
 import 'package:final_project_team02/holder/MyPage/user/components/user_tab_bar_closet.dart';
 import 'package:final_project_team02/holder/MyPage/user/components/user_tab_bar_grid.dart';
 import 'package:final_project_team02/holder/MyPage/user/components/user_tab_var.dart';
@@ -37,7 +37,7 @@ class UserMyPageBottom extends StatelessWidget {
             ),
           ],
         ),
-        UserApplyButton()
+        UserMyPageApplyButton()
       ],
     );
   }
